+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # 1.0.0
-- Implemenado suporte completo a API V1 (NFS-e)
+- Implementado suporte completo a API V1 (NFS-e)
 # 0.1.0
 - WIP (ainda não disponível para uso em produção)