+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# 1.0.1
+- Suportando valores BigDecimal para tipo :decimal
 # 1.0.0
 - Implementado suporte completo a API V1 (NFS-e)
 # 0.1.0